requirements.txt
ollama
python-dotenv
python-telegram-bot>=21,<22
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9
markdown>=3.5
